# backend/tests/fixtures/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI运营系统 API",
    "version": "1.0.0",
    "description": "AI运营系统 API 文档 - 包含价格监控和聊天功能"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/products": {
      "post": {
        "summary": "创建产品",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "platform": {
                    "type": "string",
                    "enum": [
                      "amazon",
                      "walmart",
                      "aliexpress",
                      "ebay",
                      "other"
                    ]
                  },
                  "productUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "asin": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "brand": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "category": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "currentPrice": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3,
                    "default": "USD"
                  },
                  "isMonitoring": {
                    "type": "boolean",
                    "default": false
                  },
                  "monitorType": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "scheduled"
                    ]
                  },
                  "checkInterval": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 168,
                    "default": 24
                  },
                  "userId": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                },
                "required": [
                  "platform",
                  "productUrl",
                  "asin",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "产品创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "platform": {
                      "type": "string",
                      "enum": [
                        "amazon",
                        "walmart",
                        "aliexpress",
                        "ebay",
                        "other"
                      ]
                    },
                    "productUrl": {
                      "type": "string"
                    },
                    "asin": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "currentPrice": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string"
                    },
                    "isMonitoring": {
                      "type": "boolean"
                    },
                    "monitorType": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkInterval": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "number"
                    },
                    "updatedAt": {
                      "type": "number"
                    },
                    "lastCheckedAt": {
                      "type": "number",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "platform",
                    "productUrl",
                    "asin",
                    "title",
                    "currentPrice",
                    "currency",
                    "isMonitoring",
                    "monitorType",
                    "checkInterval",
                    "userId",
                    "createdAt",
                    "lastCheckedAt",
                    "metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "获取产品列表",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "platform",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "monitoring",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "产品列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "platform": {
                            "type": "string",
                            "enum": [
                              "amazon",
                              "walmart",
                              "aliexpress",
                              "ebay",
                              "other"
                            ]
                          },
                          "productUrl": {
                            "type": "string"
                          },
                          "asin": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "brand": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "currentPrice": {
                            "type": "number",
                            "nullable": true
                          },
                          "currency": {
                            "type": "string"
                          },
                          "isMonitoring": {
                            "type": "boolean"
                          },
                          "monitorType": {
                            "type": "string",
                            "nullable": true
                          },
                          "checkInterval": {
                            "type": "number"
                          },
                          "userId": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "number"
                          },
                          "updatedAt": {
                            "type": "number"
                          },
                          "lastCheckedAt": {
                            "type": "number",
                            "nullable": true
                          },
                          "metadata": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "platform",
                          "productUrl",
                          "asin",
                          "title",
                          "currentPrice",
                          "currency",
                          "isMonitoring",
                          "monitorType",
                          "checkInterval",
                          "userId",
                          "createdAt",
                          "lastCheckedAt",
                          "metadata"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "获取产品详情",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "产品详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "platform": {
                      "type": "string",
                      "enum": [
                        "amazon",
                        "walmart",
                        "aliexpress",
                        "ebay",
                        "other"
                      ]
                    },
                    "productUrl": {
                      "type": "string"
                    },
                    "asin": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "currentPrice": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string"
                    },
                    "isMonitoring": {
                      "type": "boolean"
                    },
                    "monitorType": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkInterval": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "number"
                    },
                    "updatedAt": {
                      "type": "number"
                    },
                    "lastCheckedAt": {
                      "type": "number",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "platform",
                    "productUrl",
                    "asin",
                    "title",
                    "currentPrice",
                    "currency",
                    "isMonitoring",
                    "monitorType",
                    "checkInterval",
                    "userId",
                    "createdAt",
                    "lastCheckedAt",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "产品未找到",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "更新产品",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "platform": {
                    "type": "string",
                    "enum": [
                      "amazon",
                      "walmart",
                      "aliexpress",
                      "ebay",
                      "other"
                    ]
                  },
                  "productUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "asin": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "brand": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "category": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "currentPrice": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3,
                    "default": "USD"
                  },
                  "isMonitoring": {
                    "type": "boolean",
                    "default": false
                  },
                  "monitorType": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "scheduled"
                    ]
                  },
                  "checkInterval": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 168,
                    "default": 24
                  },
                  "userId": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "产品更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "platform": {
                      "type": "string",
                      "enum": [
                        "amazon",
                        "walmart",
                        "aliexpress",
                        "ebay",
                        "other"
                      ]
                    },
                    "productUrl": {
                      "type": "string"
                    },
                    "asin": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "brand": {
                      "type": "string"
                    },
                    "category": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "currentPrice": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string"
                    },
                    "isMonitoring": {
                      "type": "boolean"
                    },
                    "monitorType": {
                      "type": "string",
                      "nullable": true
                    },
                    "checkInterval": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "number"
                    },
                    "updatedAt": {
                      "type": "number"
                    },
                    "lastCheckedAt": {
                      "type": "number",
                      "nullable": true
                    },
                    "metadata": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "platform",
                    "productUrl",
                    "asin",
                    "title",
                    "currentPrice",
                    "currency",
                    "isMonitoring",
                    "monitorType",
                    "checkInterval",
                    "userId",
                    "createdAt",
                    "lastCheckedAt",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "产品未找到"
          }
        }
      },
      "delete": {
        "summary": "删除产品",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "产品删除成功"
          },
          "404": {
            "description": "产品未找到"
          }
        }
      }
    },
...
  }
}